Rename the file as well
